refactor(client): organize settings components into themed subfolders
The flat settings.-components/ folder had grown to ~30 files all consumed
by the single /settings route, making it hard to navigate and letting a
586-line TagGroupsAdmin accrete four inline subcomponents and six
mutations. Group the folder by the route's own tabs so structure mirrors
the UI, and decompose the one component that had outgrown a single file.

- Split into 8 tab-themed subfolders (tasks, routineTemplates, domains,
  resources, dashboard, connections, display, advanced), each with a
  public-surface-only index.ts barrel; the root barrel re-exports from
  them with identical names so settings.tsx is the only consumer touched.
- Decompose TagGroupsAdmin into a slim orchestrator plus extracted row
  components (GroupEditRow/GroupDisplayRow/TagDisplayRow/TagEditRow), a
  useTagGroupsAdmin hook, and a pure -tagDrafts module (with unit test).
- Extract the GoogleCalendar colour swatches into a FeedColorPicker leaf.
- Give every component a render-only Storybook story, including the
  Resource Settings section newly added on master in #498.
