+++ packages/client/src/routes/settings.-components/advanced/index.ts
@@ -1,0 +1,2 @@
+// Public surface of the Advanced tab.
+export { DataToolsSection } from "./-DataToolsSection";
