+++ packages/client/src/routes/settings.-components/connections/index.ts
@@ -1,0 +1,6 @@
+// Public surface of the Third Party Connections tab. -IntegrationKeySection is
+// the shared base for the Readwise/Todoist wrappers and -FeedColorPicker is
+// internal to -GoogleCalendarSection; both are intentionally not re-exported.
+export { ReadwiseSection } from "./-ReadwiseSection";
+export { TodoistSection } from "./-TodoistSection";
+export { GoogleCalendarSection } from "./-GoogleCalendarSection";
