+++ packages/client/src/routes/settings.-components/dashboard/index.ts
@@ -1,0 +1,3 @@
+// Public surface of the Dashboard Settings tab. The -useDashboardLayouts hook is
+// internal to the section and intentionally not re-exported.
+export { DashboardLayoutsSection } from "./-DashboardLayoutsSection";
