+++ packages/client/src/routes/settings.-components/display/index.ts
@@ -1,0 +1,2 @@
+// Public surface of the Display Settings tab.
+export { ThemeSection } from "./-ThemeSection";
