+++ packages/client/src/routes/settings.-components/domains/index.ts
@@ -1,0 +1,3 @@
+// Public surface of the Domains Settings tab. The -useFocusedDomains hook is
+// internal to the section and intentionally not re-exported.
+export { FocusedDomainsSection } from "./-FocusedDomainsSection";
