+++ packages/client/src/routes/settings.-components/resources/-ResourceHintTemplatesSection.tsx
