+++ packages/client/src/routes/settings.-components/resources/index.ts
@@ -1,0 +1,2 @@
+// Public surface of the Resource Settings tab.
+export { ResourceHintTemplatesSection } from "./-ResourceHintTemplatesSection";
