+++ packages/client/src/routes/settings.-components/routineTemplates/index.ts
@@ -1,0 +1,5 @@
+// Public surface of the Routine Settings tab. The edit modals
+// (-DailyCriteriaTemplateEditModal, -RoutineTemplateEditModal) are internal to
+// their sections and intentionally not re-exported.
+export { CriteriaTemplatesSection } from "./-CriteriaTemplatesSection";
+export { RoutineTemplatesSection } from "./-RoutineTemplatesSection";
