+++ packages/client/src/routes/settings.-components/tasks/index.ts
@@ -1,0 +1,6 @@
+// Public surface of the Task Settings tab. The edit rows (-TaskTypeEditRow,
+// -GroupEditRow, -GroupDisplayRow, -TagDisplayRow, -TagEditRow), the
+// -useTagGroupsAdmin hook, and the -tagDrafts helper are internal to the two
+// sections and intentionally not re-exported.
+export { TaskTypesSection } from "./-TaskTypesSection";
+export { TagGroupsAdmin } from "./-TagGroupsAdmin";
